fix: fix home link on taxonomy pages
When the current page is not the home page (there is a `.Parent`), the
"back" link is displayed. When terms lists are disabled, taxonomy pages
might have the terms list as parent (which does not exist).
The solution here is to check if .RelPermalink is set, and if not use
the home page as destination.